app_api/requirements.txt
python-dotenv == 1.1.0

pydantic == 2.10.6

pymongo == 4.11.3

loguru == 0.7.3

fastapi == 0.115.12

uvicorn == 0.34.0

pika == 1.3.2